Navigation component Added Outlet route own directory
