+++ src/routes/navigation/navigation.component.jsx
@@ -1,0 +1,15 @@
+import { Fragment } from "react";
+import { Outlet } from "react-router-dom";
+
+const Navigation = () => {
+  return (
+    <Fragment>
+      <div className="navigation">
+        <h1>I am the Navigation Bar</h1>
+      </div>
+      <Outlet />
+    </Fragment>
+  );
+};
+
+export default Navigation;
